Rename SubmitButton to VueButton
It's not always used to submit things!
